+++ lib/src/connection.dart
@@ -80,7 +80,6 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
   }
 
   final StreamController<Notification> _notifications = new StreamController<Notification>.broadcast();
-  // Add flag for debugging that captures stack trace prior to execution
 
   /// Hostname of database this connection refers to.
   String host;
@@ -109,7 +108,12 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
   /// The processID of this backend.
   int processID;
 
-  /// The notifications from the database
+  /// Stream of notification from the database.
+  ///
+  /// Listen to this [Stream] to receive events from PostgreSQL NOTIFY commands.
+  ///
+  /// To determine whether or not the NOTIFY came from this instance, compare [processID]
+  /// to [Notification.processID].
   Stream<Notification> get notifications => _notifications.stream;
 
   /// Whether or not this connection is open or not.
@@ -189,9 +193,9 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
 
     await _socket?.close();
 
-    await _notifications.close();
-
     _cancelCurrentQueries();
+
+    return _cleanup();
   }
 
   /// Executes a query on this connection.
@@ -304,6 +308,7 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
     _socket?.destroy();
 
     _cancelCurrentQueries();
+    _cleanup();
     throw new PostgreSQLException(
         "Timed out trying to connect to database postgres://$host:$port/$databaseName.");
   }
@@ -328,6 +333,7 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
   }
 
   void _cancelCurrentQueries([Object error, StackTrace stackTrace]) {
+    error ??= "Cancelled";
     var queries = _queryQueue;
     _queryQueue = [];
 
@@ -386,12 +392,14 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
     _socket.destroy();
 
     _cancelCurrentQueries(error, stack);
+    _cleanup();
   }
 
   void _handleSocketClosed() {
     _connectionState = new _PostgreSQLConnectionStateClosed();
 
     _cancelCurrentQueries();
+    _cleanup();
   }
 
   Future<Socket> _upgradeSocketToSSL(Socket originalSocket,
@@ -459,10 +467,31 @@ class PostgreSQLConnection implements PostgreSQLExecutionContext {
 
     return string;
   }
+
+  Future _cleanup() async {
+    await _notifications.close();
+  }
 }
 
 class _TransactionRollbackException implements Exception {
   _TransactionRollbackException(this.reason);
 
   String reason;
+}
+
+/// Represents a notification from PostgreSQL.
+///
+/// Instances of this type are created and sent via [PostgreSQLConnection.notifications].
+class Notification {
+  /// Creates an instance of this type.
+  Notification(this.processID, this.channel, this.payload);
+
+  /// The backend ID from which the notification was generated.
+  final int processID;
+
+  /// The name of the PostgreSQL channel that this notification occurred on.
+  final String channel;
+
+  /// An optional data payload accompanying this notification.
+  final String payload;
 }